Negative numcpus make taskq stall
Handle limited CPU VMs with 1 or 2 cores. The cpus
variable would go to -1 and taskq would stall forever.

+++ module/os/macos/zfs/vdev_disk.c
@@ -887,7 +887,7 @@ vdev_disk_init(void)
 	 * Apple Silicon we deflate further.
 	 */
 
-	const int cpus = MAX(1, max_ncpus - num_ecores - 2);
+	const int cpus = MAX(1, (int)(max_ncpus - num_ecores - 2));
 
 	/*
 	 * Keep vdev_disk_taskq_stack as in-order as we can, and use
